PockerGame.Player, PockerGame.Dealer 내부 클래스 외부로 분리
- 단, 프로퍼티 cards의 setter의 접근 범위라 fileprivate이기 때문에 Player, Dealer 클래스를 같은  파일 내에 두었음
