+++ PokerGameApp/PokerGameApp/Player.swift
@@ -1,0 +1,29 @@
+import Foundation
+
+class Player: CustomStringConvertible{
+    var name: String
+    fileprivate (set) var cards: [Card] = []
+    var description: String{
+        return "\(name) \(cards)"
+    }
+    
+    init(_ name: String){
+        self.name = name
+    }
+    
+    func addCard(_ card: Card){
+        self.cards.append(card)
+    }
+}
+
+class Dealer: Player{
+
+    init() {
+        super.init("딜러")
+    }
+    
+    func handOutCard(stud: Int)-> Card?{
+        guard let card = self.cards.popLast() else { return nil }
+        return card
+    }
+}
